refactor: 예외 발생 시 ResponseEntity으로 반환하도록 수정 #85
